requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
dill==0.2.7.1
Django==1.11.7
future==0.16.0
google-api-core==0.1.1
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-vision==0.28.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
grpcio==1.7.0
idna==2.6
ply==3.8
protobuf==3.5.0.post1
pyasn1==0.4.1
pyasn1-modules==0.2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
